refactor(internal): extract shared helpers into src/_internal
Consolidates duplicated `isPlainObject`, `UNSAFE_KEYS`, `safeAssign`, and
`isNumericSegment`/`isNumericSlice` into a single internal module. 12 object
utilities updated to import from `_internal/*` instead of carrying private
copies. Closes #162.

- Unified `isPlainObject` accepts `Object.create(null)` AND objects whose
  prototype chain ultimately resolves to `Object.prototype` — strict
  superset of the previous (inconsistent) per-file variants.
- `safeAssign` kept (used only by pick), exported for future use.
- `_internal` deliberately omitted from package.json exports.
- Public API unchanged. All per-utility size budgets still pass.
