+++ src/_internal/index.ts
@@ -1,0 +1,4 @@
+export { isNumericSegment, isNumericSlice } from "./is-numeric-segment.js";
+export { isPlainObject } from "./is-plain-object.js";
+export { safeAssign } from "./safe-assign.js";
+export { isUnsafeKey, UNSAFE_KEYS } from "./unsafe-keys.js";
