+++ src/_internal/is-plain-object.ts
@@ -1,0 +1,8 @@
+function isPlainObject(value: unknown): value is Record<string, unknown> {
+  if (typeof value !== "object" || value === null) return false;
+  const proto = Object.getPrototypeOf(value);
+  if (proto === null || proto === Object.prototype) return true;
+  return (value as { constructor?: unknown }).constructor === Object;
+}
+
+export { isPlainObject };
