+++ src/_internal/safe-assign.ts
@@ -1,0 +1,14 @@
+function safeAssign(
+  tgt: Record<string, unknown>,
+  key: string,
+  value: unknown,
+): void {
+  Object.defineProperty(tgt, key, {
+    value,
+    writable: true,
+    enumerable: true,
+    configurable: true,
+  });
+}
+
+export { safeAssign };
